Move the method to build flow into flow file
