+++ src/lib.rs
@@ -1,6 +1,6 @@
 use std::collections::HashSet;
 
-use entry::Flow;
+use flow::Flow;
 use pathfinder_geometry::transform2d::Transform2F;
 use pdf::{backend::Backend, object::{Page, Resolve}, PdfError};
 use pdf_render::{tracer::{TraceCache, Tracer, DrawItem}, Fill, render_pattern, render_page, FillMode, font::OutlineBuilder};
@@ -8,7 +8,8 @@ use pdf_render::{tracer::{TraceCache, Tracer, DrawItem}, Fill, render_pattern, r
 mod tree;
 mod util;
 mod text;
-pub mod entry;
+mod classify;
+pub mod flow;
 
 pub fn run<B: Backend>(file: &pdf::file::CachedFile<B>, page: &Page, resolve: &impl Resolve, transform: Transform2F) -> Result<Flow, PdfError> {
     let mut cache = TraceCache::new(OutlineBuilder::default());
@@ -16,11 +17,14 @@ pub fn run<B: Backend>(file: &pdf::file::CachedFile<B>, page: &Page, resolve: &i
     let mut clip_paths = vec![];
     let mut tracer = Tracer::new(&mut cache, &mut clip_paths);
 
+    // The tracer backend can be used to get text, pattern, image, etc.
+    // We will use text and pattern to do further text processing.
     render_page(&mut tracer, resolve, &page, transform)?;
 
     let bbox = tracer.view_box();
 
     let items: Vec<DrawItem<OutlineBuilder>> = tracer.finish();
+    //Get patterns which may have lines and texts inside.
     let mut patterns = HashSet::new();
     for item in items.iter() {
         if let DrawItem::Vector(ref v) = item {
@@ -35,6 +39,7 @@ pub fn run<B: Backend>(file: &pdf::file::CachedFile<B>, page: &Page, resolve: &i
 
     let mut spans = vec![];
     let mut lines = vec![];
+
     let mut visit_item = |item| {
         match item {
             DrawItem::Text(t, _) if bbox.intersects(t.rect) => {
@@ -61,6 +66,7 @@ pub fn run<B: Backend>(file: &pdf::file::CachedFile<B>, page: &Page, resolve: &i
         }
     };
 
+    // Analyze patterns to get lines and texts.
     for &p in patterns.iter() {
         let pattern = match resolve.get(p) {
             Ok(p) => p,
@@ -78,13 +84,23 @@ pub fn run<B: Backend>(file: &pdf::file::CachedFile<B>, page: &Page, resolve: &i
         }
     }
 
+    // After this loop, all the text and lines are ready
     for item in items {
         visit_item(item);
     }
-    
+  
+    spans.sort_unstable_by(|a, b| a.rect.min_y().partial_cmp(&b.rect.min_y()).unwrap());
+
+    spans.sort_unstable_by(|a, b| a.rect.min_x().partial_cmp(&b.rect.min_x()).unwrap());
+
+    for s in spans.iter().map(|s|s.text.as_str()) {
+        println!(":{}", s)
+    }
+
     let root = tree::build(&spans, bbox, &lines);
-    // dbg!(&root);
+
     let mut flow = Flow::new();
-    tree::items(&mut flow, &spans, &root, bbox.min_x());
+    flow::build(&mut flow, &spans, &root, bbox.min_x());
+
     Ok(flow)
 }
